Added 2min timer for ownership

+++ src/map/src/items/inventory.cpp
@@ -267,9 +267,19 @@ void Items::drop_item(EntitySystem& entitySystem, RoseCommon::Entity item, float
     entitySystem.add_component(item, std::move(bi));
 
     entitySystem.update_position(item, x, y);
-    
+ 
+    entitySystem.add_timer(2min, [item](EntitySystem& entitySystem) {
+        if (entitySystem.has_component<Component::Owner>(item)) {
+            entitySystem.remove_component<Component::Owner>(item);
+            auto& basic = entitySystem.get_component<Component::BasicInfo>(item);
+            basic.teamId = basic.id;
+        }
+    });
+
     entitySystem.add_timer(5min, [item](EntitySystem& entitySystem) {
-        entitySystem.delete_entity(item);
+        if (entitySystem.has_component<Component::Position>(item)) {
+            entitySystem.delete_entity(item);
+        }
     });
 }
